修复anchor使用中页面控制台显示类型转换错误 #230

+++ src/TDesign/Components/Anchor/TAnchor.cs
@@ -50,7 +50,7 @@ namespace TDesign
         /// </summary>
         /// <param name="index">scrollTop 位置</param>
         [JSInvokable]
-        public async Task OnScrollAnchorChangeAsync(double index)
+        public async Task OnScrollAnchorChangeAsync(int index)
         {
             if (!ClickLoad)
             {
